remove 32-bit compiler warning
git-svn-id: https://qthreads.googlecode.com/svn/trunk@1687 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ src/qt_c_hash.c
@@ -146,7 +146,7 @@ static inline void **qt_hash_internal_find(
 {
     const hash_entry *z = h->entries;
     const uint64_t mask = h->mask;
-    const uint64_t hashed = qt_hashword((uint64_t) key);
+    const uint64_t hashed = qt_hashword((uint64_t)(uintptr_t) key);
 
     uint64_t bucket = hashed & mask;
     uint64_t step, quit;
